Remove Auto Registration
The initial idea was to automatically register defined collectors in the default
registry because def's already have state and it would be easier to use. Because
I don't manage to support clean reloading of namespaces, I removed the auto
registration.

Now one is supposed to register all collectors in a registry using libs like
component, mount or integrant.